Fixed the statistical functions
