+++ openquake/commonlib/calculators/calc.py
@@ -203,55 +203,3 @@ def compute_hazard_maps(curves, imls, poes):
 
         result.append(hmap_val)
     return numpy.array(result).transpose()
-
-
-def mean_curve(curves, weights=None):
-    weights = weights or [1. / len(curves)] * len(curves)
-    return numpy.average(curves, axis=0, weights=weights)
-
-
-def quantile_curve(curves, quantile, weights=None):
-    """
-    Compute the weighted quantile aggregate of a set of curves
-    when using the logic tree end-branch enumeration approach, or just the
-    standard quantile when using the sampling approach.
-
-    :param curves:
-        2D array-like of curve PoEs. Each row represents the PoEs for a single
-        curve
-    :param quantile:
-        Quantile value to calculate. Should in the range [0.0, 1.0].
-    :param weights:
-        Array-like of weights, 1 for each input curve, or None
-    :returns:
-        A numpy array representing the quantile aggregate
-    """
-    if weights is None:
-        # this implementation is an alternative to
-        # numpy.array(mstats.mquantiles(curves, prob=quantile, axis=0))[0]
-        # more or less copied from the scipy mquantiles function, just special
-        # cased for what we need (and a lot faster)
-        arr = numpy.array(curves)
-        p = numpy.array(quantile)
-        m = 0.4 + p * 0.2
-        n = len(arr)
-        aleph = n * p + m
-        k = numpy.floor(aleph.clip(1, n - 1)).astype(int)
-        gamma = (aleph - k).clip(0, 1)
-        data = numpy.sort(arr, axis=0).transpose()
-        return (1.0 - gamma) * data[:, k - 1] + gamma * data[:, k]
-
-    # Each curve needs to be associated with a weight
-    assert len(weights) == len(curves)
-    weights = numpy.array(weights, dtype=numpy.float64)
-
-    result_curve = []
-    np_curves = numpy.array(curves)
-    np_weights = numpy.array(weights)
-    for poes in np_curves.transpose():
-        sorted_poe_idxs = numpy.argsort(poes)
-        sorted_weights = np_weights[sorted_poe_idxs]
-        sorted_poes = poes[sorted_poe_idxs]
-        cum_weights = numpy.cumsum(sorted_weights)
-        result_curve.append(numpy.interp(quantile, cum_weights, sorted_poes))
-    return numpy.array(result_curve)
